% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{as.lib.list}
\alias{as.lib.list}
\title{Convert a List of Data Frames to a Data Library}
\usage{
\method{as.lib}{list}(x, path)
}
\description{
Convert a List of Data Frames to a Data Library
}
\seealso{
Other lib: 
\code{\link{as.lib}()},
\code{\link{is.lib}()},
\code{\link{lib.append}()},
\code{\link{lib.copy}()},
\code{\link{lib.create}()},
\code{\link{lib.delete}()},
\code{\link{lib.info}()},
\code{\link{lib.path}()},
\code{\link{lib.remove}()},
\code{\link{lib.size}()},
\code{\link{lib.write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
