% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{as.lib}
\alias{as.lib}
\title{Generic Casting Method for Data Libraries}
\usage{
as.lib(x)
}
\arguments{
\item{x}{The object to cast.}
}
\value{
A library object, created using the information in the 
input object.
}
\description{
A generic method for casting objects to
a data library.  Individual objects will inherit from this function.
}
\seealso{
For class-specific methods, see 
\code{\link{as.lib.list}}.

Other lib: 
\code{\link{as.lib.list}()},
\code{\link{is.lib}()},
\code{\link{lib.append}()},
\code{\link{lib.copy}()},
\code{\link{lib.create}()},
\code{\link{lib.delete}()},
\code{\link{lib.info}()},
\code{\link{lib.path}()},
\code{\link{lib.remove}()},
\code{\link{lib.size}()},
\code{\link{lib.write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
