% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Create a Data Dictionary}
\usage{
dictionary(x)
}
\arguments{
\item{x}{The input library, data frame, or tibble.}
}
\description{
A function to create a data dictionary for a data frame,
a tibble, or a data library.  The function will generate a tibble of 
information about the data.  The tibble will contain the following columns:
}
