% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsarray.R
\encoding{UTF-8}
\name{dsarray}
\alias{dsarray}
\title{Create a Datastep Array}
\usage{
dsarray(...)
}
\arguments{
\item{...}{Column names to include as part of the datastep array.
Separate column identifiers with commas.}
}
\value{
The datastep array object.
}
\description{
A function to capture a set of column names with
which to create a datastep array.  The datastep array can be used inside
the \code{\link{datastep}} to loop over columns.
}
\examples{
library(libr)
}
\seealso{
\code{\link{libname}} to create a data library, and
\code{\link{dictionary}} for generating a data dictionary

Other datastep: 
\code{\link{[.dsarray}()},
\code{\link{datastep}()},
\code{\link{dsattr}()},
\code{\link{length.dsarray}()}
}
\concept{datastep}
