% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsattr.R
\encoding{UTF-8}
\name{dsattr}
\alias{dsattr}
\title{Assign Datastep Variable Attributes}
\usage{
dsattr(
  class = NULL,
  label = NULL,
  description = NULL,
  width = NULL,
  format = NULL,
  justify = NULL,
  ...
)
}
\arguments{
\item{class}{The desired class of the column. Valid values are 
'character', 'numeric', 'integer', 'logical', and 'complex'.}

\item{label}{The label to associate with this column.  Accepts
any string value.}

\item{description}{A description for this column.  Accepts any string
value.}

\item{width}{The desired width for the column.  Works only on character
columns. Applying a width to a column will make it fixed-width.}

\item{format}{The format associated with this column.  See the 
\code{\link[fmtr]{fmtr}} package for more information about formatting.}

\item{justify}{The desired justification for the column.  This parameter
it normally used only for fixed-width, character columns.}

\item{...}{Any other attributes you wish to assign to this column.}
}
\value{
The import specifications object.
}
\description{
A function to capture a set of column names with
which to create a datastep array.  The datastep array can be used inside
the \code{\link{datastep}} to loop over columns.
}
\examples{
library(libr)
}
\seealso{
\code{\link{libname}} to create a data library, and
\code{\link{dictionary}} for generating a data dictionary

Other datastep: 
\code{\link{[.dsarray}()},
\code{\link{datastep}()},
\code{\link{dsarray}()},
\code{\link{length.dsarray}()}
}
\concept{datastep}
