% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%eq\%}
\alias{\%eq\%}
\title{Check identity of two objects}
\usage{
x1 \%eq\% x2
}
\arguments{
\item{x1}{The first object to compare}

\item{x2}{The second object to compare}
}
\value{
A TRUE or FALSE value depending on whether the objects are equal.
}
\description{
The \code{\%eq\%} function provides a more human-comprehensible
style of equality check.  The function allows comparing of nulls, NA values,
and atomic type values without error. The function also allows comparing
of data frames.  The function will return TRUE if all values in the 
data frames are equal, and ignores differences in attributes.
}
