% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsarray.R
\encoding{UTF-8}
\name{length.dsarray}
\alias{length.dsarray}
\title{Length function for dsarray class}
\usage{
\method{length}{dsarray}(x)
}
\arguments{
\item{x}{The dsarray object.}
}
\value{
The number of items in the specified dsarray.
}
\description{
A length function for the Datastep Array.
}
\seealso{
Other datastep: 
\code{\link{[.dsarray}()},
\code{\link{datastep}()},
\code{\link{dsarray}()},
\code{\link{dsattr}()}
}
\concept{datastep}
