% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib.path}
\alias{lib.path}
\title{Get the Path for a Data Library}
\usage{
lib.path(x)
}
\description{
Get the Path for a Data Library
}
\seealso{
Other lib: 
\code{\link{as.lib.list}()},
\code{\link{as.lib}()},
\code{\link{is.lib}()},
\code{\link{lib.append}()},
\code{\link{lib.copy}()},
\code{\link{lib.create}()},
\code{\link{lib.delete}()},
\code{\link{lib.info}()},
\code{\link{lib.remove}()},
\code{\link{lib.size}()},
\code{\link{lib.write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
