% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib.remove}
\alias{lib.remove}
\title{Remove Data from a Data Library}
\usage{
lib.remove(x)
}
\description{
Remove Data from a Data Library
}
\seealso{
Other lib: 
\code{\link{as.lib.list}()},
\code{\link{as.lib}()},
\code{\link{is.lib}()},
\code{\link{lib.append}()},
\code{\link{lib.copy}()},
\code{\link{lib.create}()},
\code{\link{lib.delete}()},
\code{\link{lib.info}()},
\code{\link{lib.path}()},
\code{\link{lib.size}()},
\code{\link{lib.write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
