% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib.write}
\alias{lib.write}
\title{Write a data library to the file system}
\usage{
lib.write(
  x,
  dir_path = getwd(),
  file_name = deparse(substitute(x, env = environment())),
  type = NULL
)
}
\arguments{
\item{x}{The format catalog to write.}

\item{dir_path}{The directory path to write the catalog to. Default is the 
current working directory.}

\item{file_name}{The name of the file to save the catalog as.  Default is
the name of the variable that contains the catalog.  The ".fcat" file
extension will be added automatically.}

\item{type}{The type of data library. Default is NULL, meaning the library
is not typed.  Valid values are "rds", "sas7bdat", "xls", "xlsx", and "csv". 
The data files will be saved as the type specified on this parameter, 
no matter from which type of file they were input.}
}
\value{
The full path of the saved format catalog.
}
\description{
The \code{lib.write} function writes the data library
to the file system.  By default, the library will be written to the 
directory for which it was defined, and each data frame will be written
in the format from which it was read.  Data frames that were not read
from a file will be saved in RDS format, unless otherwise specified.
}
\seealso{
Other lib: 
\code{\link{as.lib.list}()},
\code{\link{as.lib}()},
\code{\link{is.lib}()},
\code{\link{lib.append}()},
\code{\link{lib.copy}()},
\code{\link{lib.create}()},
\code{\link{lib.delete}()},
\code{\link{lib.info}()},
\code{\link{lib.path}()},
\code{\link{lib.remove}()},
\code{\link{lib.size}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
