% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib_add}
\alias{lib_add}
\title{Add Data to a Data Library}
\usage{
lib_add(x, ..., type = NULL, .name = NULL)
}
\arguments{
\item{x}{The library to add data to.}

\item{...}{The data frame(s) to add to the library.}

\item{type}{The type of file format to assign to this data. By default,
the function will assign the type associated with the library.  If no
type is associated with the library, the type will be assigned "rds".}

\item{.name}{The reference name to use for the data.  By default,
the name will be the variable name.  To assign a name different
from the variable name, assign a quoted name to this parameter.  If more
than one data set is being appended, assign a vector of quoted names.}
}
\description{
The \code{\link{lib_add}} function adds a data frame
to an existing data library.
}
\seealso{
Other lib: 
\code{\link{is.lib}()},
\code{\link{lib_copy}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_environment}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{lib_write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
