% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib_append}
\alias{lib_append}
\title{Append Data to a Data Library}
\usage{
lib_append(x, df)
}
\arguments{
\item{x}{The library to append data to.}

\item{df}{The data frame to append to the library.}
}
\description{
The \code{\link{lib_append}} function adds a data frame
to an existing data library.
}
\seealso{
Other lib: 
\code{\link{as.lib.list}()},
\code{\link{as.lib}()},
\code{\link{is.lib}()},
\code{\link{lib_copy}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{lib_write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
