% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib_copy}
\alias{lib_copy}
\title{Copy a Data Library}
\usage{
lib_copy(x)
}
\arguments{
\item{x}{The library to copy.}
}
\description{
Copy a Data Library
}
\seealso{
Other lib: 
\code{\link{as.lib.list}()},
\code{\link{as.lib}()},
\code{\link{is.lib}()},
\code{\link{lib_append}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{lib_write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
