% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib_create}
\alias{lib_create}
\title{Create a New Data Directory and Library}
\usage{
lib_create(directory_path)
}
\arguments{
\item{directory_path}{The path of the directory to create.}
}
\value{
The new data library.
}
\description{
The \code{\link{lib_create}} function creates a new 
data directory if it does not exist, and creates a new library
assigned to that directory.
}
\seealso{
Other lib: 
\code{\link{is.lib}()},
\code{\link{lib_append}()},
\code{\link{lib_copy}()},
\code{\link{lib_delete}()},
\code{\link{lib_environment}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{lib_write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
