% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib_delete}
\alias{lib_delete}
\title{Delete a Data Library}
\usage{
lib_delete(x)
}
\arguments{
\item{x}{The data library to delete.}
}
\description{
The \code{lib_delete} function deletes a data library from
the file system.  All data files associated with the library will be deleted.
If other files exist in the library directory, they will not be affected
by the delete operation.  The directory that contains the data will also
not be affected by the delete operation.  To delete the data directory, 
use the \code{\link[base]{unlink}} function.
}
\seealso{
Other lib: 
\code{\link{as.lib.list}()},
\code{\link{as.lib}()},
\code{\link{is.lib}()},
\code{\link{lib_append}()},
\code{\link{lib_copy}()},
\code{\link{lib_create}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{lib_write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
