% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib_environment}
\alias{lib_environment}
\title{Set the Environment for the Library Functions}
\usage{
lib_environment(env = NULL)
}
\arguments{
\item{env}{The environment to use.}
}
\value{
The new environment.
}
\description{
The \code{lib_environment} function sets the working environment
for the \strong{libr} package.  By default, the environment is set to the 
current global environment.  If an alternative environment is desired, a 
user may set it with this function.
}
\seealso{
Other lib: 
\code{\link{is.lib}()},
\code{\link{lib_append}()},
\code{\link{lib_copy}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{lib_write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
