% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib_unload}
\alias{lib_unload}
\title{Unload a Library from the Workspace}
\usage{
lib_unload(x, sync = TRUE, .name = NULL)
}
\arguments{
\item{x}{The data library to unload.}

\item{sync}{Whether to sync the workspace with the library list before
it is unloaded.  If you want to unload the workspace without saving the 
workspace data, set this parameter to false.}

\item{.name}{The name of the library to unload, if the name is different
than the variable name.  Used internally.}
}
\value{
The unloaded data library.
}
\description{
The \code{lib_unload} function unloads a data library from
the workspace environment.  The unload function does not delete the data 
or the remove the library.  It simply removes the data frames from working 
memory.
}
\seealso{
\code{\link{lib_load}} to load the library.

Other lib: 
\code{\link{is.lib}()},
\code{\link{lib_add}()},
\code{\link{lib_copy}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_environment}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
