% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastep.R
\name{output}
\alias{output}
\title{Outputs an observation from a datastep}
\usage{
output()
}
\value{
Observation is marked with a output flag.  No return value.
}
\description{
The \code{output} function will output an observation
from a datastep.  The function takes no parameters.  To use 
the function, simply call it on the rows you want to output.  Typically
it is called within a conditional.  The output function is interesting
in that you can output multiple rows for the same input observation.
}
\examples{
#' # Example 1: Output all cars that are 4 cylinder 
df <- datastep(mtcars, 
               keep = c("mpg", "cyl", "disp"), {
                 
  if (cyl == 4)
    output()
                 
})

df
#     mpg cyl  disp
# 1  22.8   4 108.0
# 2  24.4   4 146.7
# 3  22.8   4 140.8
# 4  32.4   4  78.7
# 5  30.4   4  75.7
# 6  33.9   4  71.1
# 7  21.5   4 120.1
# 8  27.3   4  79.0
# 9  26.0   4 120.3
# 10 30.4   4  95.1
# 11 21.4   4 121.0

# Example 2: Output two rows for each 6 cylinder car
df <- datastep(mtcars, 
               keep = c("mpg", "cyl", "disp", "seq"), {
                 
  if (cyl == 6) {
    seq <- 1
    output()
    seq <- 2
    output()
  }
                 
})

df
#     mpg cyl  disp seq
# 1  21.0   6 160.0   1
# 2  21.0   6 160.0   2
# 3  21.0   6 160.0   1
# 4  21.0   6 160.0   2
# 5  21.4   6 258.0   1
# 6  21.4   6 258.0   2
# 7  18.1   6 225.0   1
# 8  18.1   6 225.0   2
# 9  19.2   6 167.6   1
# 10 19.2   6 167.6   2
# 11 17.8   6 167.6   1
# 12 17.8   6 167.6   2
# 13 19.7   6 145.0   1
# 14 19.7   6 145.0   2
}
\seealso{
The \code{\link{datastep}} function.
}
