% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{print.lib}
\alias{print.lib}
\title{Print a data library}
\usage{
\method{print}{lib}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{The library to print.}

\item{...}{Any follow-on parameters.}

\item{verbose}{Whether or not to print the library in verbose style.
By default, the parameter is FALSE, meaning to print in summary style.}
}
\value{
The object, invisibly.
}
\description{
A class-specific instance of the \code{print} function for 
data libraries.  The function prints the library in a summary manner.  
Use \code{verbose = TRUE} to print the catalog as a list.
}
\examples{
# Create temp directory
tmp <- tempdir()

# Save some data to temp directory
saveRDS(iris, file.path(tmp, "iris.rds"))
saveRDS(ToothGrowth, file.path(tmp, "ToothGrowth.rds"))
saveRDS(PlantGrowth, file.path(tmp, "PlantGrowth.rds"))

# Create format catalog
libname(dat, tmp)

# Print library summary 
print(dat)
}
\seealso{
Other lib: 
\code{\link{as.lib.list}()},
\code{\link{as.lib}()},
\code{\link{is.lib}()},
\code{\link{lib_append}()},
\code{\link{lib_copy}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{lib_write}()},
\code{\link{libname}()}
}
\concept{lib}
