% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{print.specs}
\alias{print.specs}
\title{Print import specifications}
\usage{
\method{print}{specs}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{The specifications to print.}

\item{...}{Any follow-on parameters to the print function.}

\item{verbose}{Whether or not to print the specifications in verbose style.
By default, the parameter is FALSE, meaning to print in summary style.}
}
\value{
The specification object, invisibly.
}
\description{
A function to print the import specification collection.
}
\seealso{
Other specs: 
\code{\link{import_spec}()},
\code{\link{read.specs}()},
\code{\link{specs}()},
\code{\link{write.specs}()}
}
\concept{specs}
