% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{read.specs}
\alias{read.specs}
\title{Read input specs from the file system}
\usage{
read.specs(file_path)
}
\arguments{
\item{file_path}{The full or relative path to the file system.  This 
parameter is required. If the file_path does not contain the '.specs' file
extension, the function will add it.}
}
\value{
The specifications object.
}
\description{
A function to read input specifications from the file system.
The function accepts a full or relative path to the spec file, and returns
the specs as an object.
}
