% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsarray.R
\encoding{UTF-8}
\name{[.dsarray}
\alias{[.dsarray}
\title{Indexer for Datastep Array}
\usage{
\method{[}{dsarray}(x, i = NULL)
}
\arguments{
\item{x}{The dsarray object.}

\item{i}{The index of the datastep array item to return a value for.
This index can be a column name or position in the array.  If no index
is supplied, a vector of all dataset values will be returned.}
}
\value{
The value of the specified column for the current row in the 
datastep.  If no index is supplied, a vector of all column values will
be returned.
}
\description{
A custom indexer for the Datastep Array.  The indexer will
return an individual row/column value for the specified column name.  
For additional details, see the \code{\link{dsarray}} function.
}
\seealso{
Other datastep: 
\code{\link{datastep}()},
\code{\link{dsarray}()},
\code{\link{length.dsarray}()}
}
\concept{datastep}
