% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{write.specs}
\alias{write.specs}
\title{Write input specs to the file system}
\usage{
write.specs(x, dir_path = getwd(), file_name = NULL)
}
\arguments{
\item{x}{A specifications object of class 'specs'.}

\item{dir_path}{A full or relative path to save the specs.}

\item{file_name}{The file name to save to specs, without a file extension.
The file extension will be added automatically.  If no file name is 
supplied, the function will use the variable name as the file name.}
}
\description{
A function to write input specifications to the file system.
The function accepts a specifications object and a full or relative
path.  The function returns the full file path.
}
