% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{as.lib.list}
\alias{as.lib.list}
\title{Convert a List of Data Frames to a Data Library}
\usage{
\method{as.lib}{list}(x, path)
}
\arguments{
\item{x}{The list to convert}

\item{path}{The path to associate with the library.}
}
\description{
Convert a List of Data Frames to a Data Library
}
\seealso{
Other lib: 
\code{\link{as.lib}()},
\code{\link{is.lib}()},
\code{\link{lib_append}()},
\code{\link{lib_copy}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_environment}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{lib_write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
