% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{as.lib}
\alias{as.lib}
\title{Generic Casting Method for Data Libraries}
\usage{
as.lib(x, path)
}
\arguments{
\item{x}{The object to cast.}

\item{path}{The path for the specified library.}
}
\value{
A library object, created using the information in the 
input object.
}
\description{
A generic method for casting objects to
a data library.  Individual objects will inherit from this function.
}
\seealso{
For class-specific methods, see 
\code{\link{as.lib.list}}.

Other lib: 
\code{\link{as.lib.list}()},
\code{\link{is.lib}()},
\code{\link{lib_append}()},
\code{\link{lib_copy}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_environment}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{lib_write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
