% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastep.R
\name{datastep}
\alias{datastep}
\title{Step through data row by row}
\usage{
datastep(
  data,
  steps,
  keep = NULL,
  drop = NULL,
  rename = NULL,
  by = NULL,
  calculate = NULL
)
}
\arguments{
\item{data}{The data to step through.}

\item{steps}{The operations to perform on the data.  This parameter is 
typically specified as a set of R statements contained within 
curly braces.}

\item{keep}{A vector of quoted variable names to keep in the output
data set. By default, all variables are kept.}

\item{drop}{A vector of quoted variable names to drop from the output
data set. By default, no variables are dropped.}

\item{rename}{A named vector of quoted variables to rename.  The current
variable name should be on the left had side of the vector name/value pair,
and the new variable name should be on the right.  The rename operation
is performed after the data step, the keep, and the drop.  Therefore, 
the data steps should use the old variable name.  By default, all variable
retain their original names.}

\item{by}{A vector of quoted variable names to use for by-group processing.
This parameter will activate the \code{first.} and \code{last.} special
variables, that indicate the first or last rows in a group.  These 
special variables are useful for conditional processing on groups.}

\item{calculate}{Steps to set up calculated variables.
Calculated variables are common performed with summary functions such as
\code{mean}, \code{median}, \code{min}, \code{max}, etc.  It is more 
efficient to set up calculated variables with the calculate parameter and then 
use those variables in the data step, rather than perform the summary
function inside the data step.}
}
\value{
The processed data frame or tibble.
}
\description{
The \code{datastep} function allows you to perform
row-wise conditional processing on a data frame or tibble. The function
contains parameters to drop, keep, or rename variables.  It also 
has the ability to perform by-group processing, and identify first
and last rows in the group.
}
