% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsarray.R
\encoding{UTF-8}
\name{length.dsarray}
\alias{length.dsarray}
\title{Length function for dsarray class}
\usage{
\method{length}{dsarray}(x)
}
\arguments{
\item{x}{The \code{\link{dsarray}} object.}
}
\value{
The number of items in the specified \code{\link{dsarray}}.
}
\description{
A length function for the data step array \code{\link{dsarray}}.
The length function can by used either inside or outside the data step.
}
\examples{
# Define datastep array
carr <- dsarray(names(mtcars))

length(carr)
# 11
}
\seealso{
Other datastep: 
\code{\link{[.dsarray}()},
\code{\link{datastep}()},
\code{\link{delete}()},
\code{\link{dsarray}()},
\code{\link{dsattr}()},
\code{\link{output}()}
}
\concept{datastep}
