% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib_copy}
\alias{lib_copy}
\title{Copy a Data Library}
\usage{
lib_copy(x, nm, directory_path)
}
\arguments{
\item{x}{The library to copy.}

\item{nm}{The unquoted variable name to hold the new library.}

\item{directory_path}{The path to copy the library to.}
}
\description{
The \code{lib_copy} function copies a data library.  The 
function accepts a library and a destination path.  If the destination 
path does not exist, the function will attempt to create it.  Note that
the copy will result in the current data in memory written to the new
destination directory.
}
\seealso{
Other lib: 
\code{\link{is.lib}()},
\code{\link{lib_add}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_environment}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{lib_write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
