% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib_info}
\alias{lib_info}
\title{Get Information on a Data Library}
\usage{
lib_info(x)
}
\arguments{
\item{x}{The data library.}
}
\value{
A data frame of information about the library.
}
\description{
The \code{lib_info} function returns a data frame of information
about each item in the data library.  That information includes the item
name, file extension, number of rows, number of columns, size in bytes, 
the last modified date.
}
\seealso{
Other lib: 
\code{\link{as.lib.list}()},
\code{\link{as.lib}()},
\code{\link{is.lib}()},
\code{\link{lib_append}()},
\code{\link{lib_copy}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{lib_write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
