% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib_load}
\alias{lib_load}
\title{Load a Library into the Global Environment}
\usage{
lib_load(x)
}
\arguments{
\item{x}{The data library to load.}
}
\value{
The loaded data library.
}
\description{
The \code{lib_load} function loads a data library into
the global environment.  The data frames will be loaded with 
<library>.<data frame> syntax.
}
\seealso{
\code{\link{lib_unload}} to unload the library.

Other lib: 
\code{\link{as.lib.list}()},
\code{\link{as.lib}()},
\code{\link{is.lib}()},
\code{\link{lib_append}()},
\code{\link{lib_copy}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_environment}()},
\code{\link{lib_info}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{lib_write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
