% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib_path}
\alias{lib_path}
\title{Get the Path for a Data Library}
\usage{
lib_path(x)
}
\arguments{
\item{x}{The data library.}
}
\description{
Get the Path for a Data Library
}
\seealso{
Other lib: 
\code{\link{as.lib.list}()},
\code{\link{as.lib}()},
\code{\link{is.lib}()},
\code{\link{lib_append}()},
\code{\link{lib_copy}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{lib_write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
