% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib_remove}
\alias{lib_remove}
\title{Remove Data from a Data Library}
\usage{
lib_remove(x, name)
}
\arguments{
\item{x}{The data library.}

\item{name}{The quoted name of the item to remove from the data library. 
For more than one name, pass a vector of quoted names.}
}
\value{
The library with the requested item removed.
}
\description{
The \code{lib_remove} function removes an item from the 
data library, and deletes the source file for that data.
}
\seealso{
Other lib: 
\code{\link{as.lib.list}()},
\code{\link{as.lib}()},
\code{\link{is.lib}()},
\code{\link{lib_append}()},
\code{\link{lib_copy}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_environment}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{lib_write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
