% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib_unload}
\alias{lib_unload}
\title{Unload a Library from the Global Environment}
\usage{
lib_unload(x)
}
\arguments{
\item{x}{The data library to unload.}
}
\value{
The unloaded data library.
}
\description{
The \code{lib_unload} function unloads a data library from
the global environment.  The unload function does not delete the data 
or the remove the library.  It simply removes the data frames from global 
memory.
}
\seealso{
\code{\link{lib_load}} to load the library.

Other lib: 
\code{\link{as.lib.list}()},
\code{\link{as.lib}()},
\code{\link{is.lib}()},
\code{\link{lib_append}()},
\code{\link{lib_copy}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_write}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
