% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib_write}
\alias{lib_write}
\title{Write a Data Library to the File System}
\usage{
lib_write(x, type = NULL)
}
\arguments{
\item{x}{The format catalog to write.}

\item{type}{The type of data library. Default is NULL, meaning the library
is not typed.  Valid values are "rds", "sas7bdat", "xls", "xlsx", "csv",
and "dat".
The data files will be saved as the type specified on this parameter, 
no matter from which type of file they were input.  If the type is
passed as a single string, all data files will be saved as that single type.
The type can also be passed as a named vector of types, with each name 
corresponding to the name of the data frame in the library.}
}
\value{
The saved data library.
}
\description{
The \code{lib_write} function writes the data library
to the file system.  By default, the library will be written to the 
directory for which it was defined, and each data frame will be written
in the format from which it was read.  Data frames that were not read
from a file will be saved in RDS format, unless otherwise specified.  To
specify a different format, set the \code{type} parameter on 
\code{lib_write}.  Setting this parameter will override any source
file types and save the data in the indicated file format.

The \strong{libr} package can write data files in csv, rds, xlsx, sas7bdat,
and a plain text file format.  Data read in xls file format will be saved 
as an xlsx.
}
\examples{
# Create temp directory
tmp <- tempdir()

# Create library
libname(dat, tmp)

# Load the empty library 
lib_load(dat)

# Add data to the library
dat.mtcars <- mtcars
dat.beaver1 <- beaver1
dat.iris <- iris

# Unload the library
lib_unload(dat)

# Write the library to the file system
lib_write(dat)

# Examine the library
dat
# library 'dat': 3 items
#- attributes: not loaded
#- path: C:\Users\User\AppData\Local\Temp\RtmpCSJ6Gc
#- items:
#     Name Extension Rows Cols   Size        LastModified
#1 beaver1       rds  114    4 4.8 Kb 2020-11-05 20:47:16
#2    iris       rds  150    5 7.3 Kb 2020-11-05 20:47:16
#3  mtcars       rds   32   11 7.3 Kb 2020-11-05 20:47:16

# Clean up
lib_delete(dat)
}
\seealso{
Other lib: 
\code{\link{is.lib}()},
\code{\link{lib_add}()},
\code{\link{lib_copy}()},
\code{\link{lib_delete}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_replace}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
