% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib_write}
\alias{lib_write}
\title{Write a data library to the file system}
\usage{
lib_write(x, type = NULL)
}
\arguments{
\item{x}{The format catalog to write.}

\item{type}{The type of data library. Default is NULL, meaning the library
is not typed.  Valid values are "rds", "sas7bdat", "xls", "xlsx", "csv",
and "dat".
The data files will be saved as the type specified on this parameter, 
no matter from which type of file they were input.  If the type is
passed as a single string, all data files will be saved as that single type.
The type can also be passed as a named vector of types, with each name 
corresponding to the name of the data frame in the library.}
}
\value{
The saved data library.
}
\description{
The \code{lib_write} function writes the data library
to the file system.  By default, the library will be written to the 
directory for which it was defined, and each data frame will be written
in the format from which it was read.  Data frames that were not read
from a file will be saved in RDS format, unless otherwise specified.  To
specify a different format, set the \code{type} parameter on 
\code{lib_write}.  Setting this parameter will override any source
file types and save the data in the specified file format.

Note that there are some limitations on file formats for writing.  The 
\strong{libr} package cannot currently write to sas7bdat file format.  The 
reason is that there are currently no easy to install and reliable
R packages for writing SAS® files.  

Likewise, there are few easily installed
package for writing to Excel.  The \strong{libr} package uses the 'xlsx' 
package for writing to Excel, and this package requires a properly installed
and compatible version of Java.  If the 'xlsx' package is installed,
\strong{libr} will attempt to use it.  If the package is not installed, 
the \code{lib_write} function will give a message, and not write the file.
}
\seealso{
Other lib: 
\code{\link{is.lib}()},
\code{\link{lib_add}()},
\code{\link{lib_copy}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_environment}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
