% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{lib_write}
\alias{lib_write}
\title{Write a data library to the file system}
\usage{
lib_write(x, type = NULL)
}
\arguments{
\item{x}{The format catalog to write.}

\item{type}{The type of data library. Default is NULL, meaning the library
is not typed.  Valid values are "rds", "sas7bdat", "xls", "xlsx", "csv",
and "dat".
The data files will be saved as the type specified on this parameter, 
no matter from which type of file they were input.  If the type is
passed as a single string, all data files will be saved as that single type.
The type can also be passed as a named vector of types, with each name 
corresponding to the name of the data frame in the library.}
}
\value{
The full path of the saved data library.
}
\description{
The \code{lib_write} function writes the data library
to the file system.  By default, the library will be written to the 
directory for which it was defined, and each data frame will be written
in the format from which it was read.  Data frames that were not read
from a file will be saved in RDS format, unless otherwise specified.  To
specify a different format, set the \code{type} parameter on 
\code{lib_write}.  Setting this parameter will override any source
file types and save the data in a consistent file format.
}
\seealso{
Other lib: 
\code{\link{as.lib.list}()},
\code{\link{as.lib}()},
\code{\link{is.lib}()},
\code{\link{lib_append}()},
\code{\link{lib_copy}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{libname}()},
\code{\link{print.lib}()}
}
\concept{lib}
