% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libname.R
\name{libname}
\alias{libname}
\title{Create a data library}
\usage{
libname(name, directory_path, filter = NULL, read_only = FALSE, pos = 1, ...)
}
\arguments{
\item{name}{The unquoted name of the library to create.  This name will 
be created as a variable in the global environment.}

\item{directory_path}{A directory path in which the data resides.  The incoming 
data can be in any file format: rds, csv, sas7bdat, etc.  The 
\code{libname} function will read any type of data into the library, 
and render as an R data frame.}

\item{filter}{One or more file extensions to filter the incoming data.  The
default value is NULL, meaning all recognized data files will be input.
Multiple filters can be passed as a vector of file extensions. 
Valid values are "rds", "sas7bdat", "xls", "xlsx", "csv", and "dat".}

\item{read_only}{Whether the library should be created as read only.
Default is FALSE.  If TRUE, the user will be restricted from
appending, removing, or writing data from the library to the file system.}

\item{pos}{The environment to load the library into.  This parameter is
used internally to the package.}

\item{...}{Follow-on parameters to the data import functions.  Which
parameters exist depend on which types of files are being imported.}
}
\value{
The library object.
}
\description{
A data library is a collection of data frames  A data 
library allows you to manage and store data frames as a unit.  The 
\code{libname} function defines the library.
}
\details{
A data library is an S3 object of class "lib".  The purpose of 
the library is to combine related data frames, and allow you to manipulate all
of them as a single object.
}
\examples{
# Create temp directory
tmp <- tempdir()

# Save some data to temp directory
saveRDS(trees, file.path(tmp, "trees.rds"))
saveRDS(lynx, file.path(tmp, "lynx.rds"))
saveRDS(beaver1, file.path(tmp, "beaver1.rds"))

# Create data library
libname(dat, tmp)

# Print library summary 
print(dat)
}
\seealso{
Other lib: 
\code{\link{as.lib.list}()},
\code{\link{as.lib}()},
\code{\link{is.lib}()},
\code{\link{lib_append}()},
\code{\link{lib_copy}()},
\code{\link{lib_create}()},
\code{\link{lib_delete}()},
\code{\link{lib_info}()},
\code{\link{lib_load}()},
\code{\link{lib_path}()},
\code{\link{lib_remove}()},
\code{\link{lib_size}()},
\code{\link{lib_unload}()},
\code{\link{lib_write}()},
\code{\link{print.lib}()}
}
\concept{lib}
