% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{specs}
\alias{specs}
\title{Create an Import Spec Collection}
\usage{
specs(col_types = NULL, na = "", trim_ws = TRUE, ...)
}
\arguments{
\item{col_types}{A vector of column data types and optional input format.
Available types are: 'guess', 'logical', 'character', 'integer', 'numeric',
'date', 'datetime', and 'time'.  The date/time data types accept an optional
input format.  To supply the input format, append it after the data type
following an equals sign, e.g.: 'date=%d%B%Y' or 'datetime=%d%m%Y %H:%M:%S'.
Default is NULL, meaning no column types are specified, and the function
should make its best guess for each column.}

\item{na}{A vector of values to be treated as NA.  For example, the 
vector \code{c('', ' ')} will cause empty strings and single blanks to 
be converted to NA values. Default is that empty strings ('') are considered
NA.}

\item{trim_ws}{Whether or not to trim white space from the input data values.
Valid values are TRUE, and FALSE.  Default is TRUE.}

\item{...}{Named input specs.  The name should correspond to the file name.}
}
\description{
A function to capture a set of import specifications for a 
directory of data files.  These specs can be used on the 
\code{\link{libname}} function to correctly assign the data types for 
file formats that are not strongly typed, such as 'csv' and 'xlsx'. The 
input specification is defined as an object so it can be stored and reused.
See the \code{\link{write.specs}} and \code{\link{read.specs}} functions
for additional information on storing specs.
}
